+++ lib/components/dropdown_menus/styles/m3e_dropdown_chip_style.dart
@@ -3,7 +3,7 @@ import 'dart:ui';
 import 'package:flutter/foundation.dart';
 import 'package:material_ui/material_ui.dart';
 
-import '../../../foundations/m3e_motion.dart';
+import '../../../foundations/foundations.dart';
 
 /// M3EDropdownChipStyle.