+++ lib/foundations/theme/m3e_theme_defaults.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/foundation.dart';
 
-import 'm3e_color_scheme.dart';
+import '../color/m3e_color_scheme.dart';
+import '../type/m3e_typography.dart';
 import 'm3e_theme_data.dart';
-import 'm3e_typography.dart';
 
 /// Assembles a complete [M3EThemeData] from core tokens and component defaults.
 M3EThemeData buildM3EThemeDefaults({